fix(docs): ruby comments use # not //
Hello Merge friends! Congrats on getting this out there, it looks like a great SDK at first blush.

Just a small nit that some comments in the README are using the wrong syntax. I do it all the time :)

Note: I am aware the repo is generated and my change may not be merged as-is, but figured a PR would still be more helpful than raising an issue.